requirements.txt
websocket-client>=1.6.0
requests>=2.31.0
colorama>=0.4.6
